statically cast param->axis to size_t.

+++ src/relay/op/tensor/transform.cc
@@ -255,7 +255,7 @@ bool ConcatenateRel(const Array<Type>& types,
       throw relay::Error("relay.concatenate requires all tensors have the same dtype");
     }
     for (size_t j = 0; j < first->shape.size(); ++j) {
-      if (j == param->axis) continue;
+      if (j == static_cast<size_t>(param->axis)) continue;
       if (reporter->AssertEQ(first->shape[j], e->shape[j])) continue;
       throw relay::Error("relay.concatenate requires all tensors have the same shape "
                          "on non-concatenating axes");
@@ -396,7 +396,7 @@ bool StackRel(const Array<Type>& types,
     CHECK_EQ(e_ndim, ndim) << "relay.stack requires all tensors have the same ndim";
     CHECK_EQ(e_dtype, dtype) << "relay.stack requires all tensors have the same dtype";
     for (size_t j = 0; j < first->shape.size(); ++j) {
-      if (j == param->axis) continue;
+      if (j == static_cast<size_t>(param->axis)) continue;
       if (reporter->AssertEQ(first->shape[j], e->shape[j])) continue;
       throw relay::Error("relay.stack requires all tensors have the same shape "
                          "on non-stacking axes");
